requirements.txt
flask
Pillow
gunicorn